detection/minos/src/requirements.txt
keras
matplotlib
modin
numpy
pandas
Pillow
scikit_learn
tensorflow
